examples/multi-kb/requirements.txt
accelerate
pandas
faiss-gpu
einops
gpt4pandas